# Do not flush NewStream header on client side for unary RPCs and streaming RPCs with requests.