If it's not client streaming, we should already have the request to be sent, so we don't flush the header.
If it is client streaming, the user may never send a request or send it any time soon, so we ask the transport to flush the header.

This change reduced one system call for flushing the headers.

Without the change:
```
Unary-Tracing-c_1-req_1B-resp_1B_noMD-12         	   10000	    103777 ns/op
Unary-Tracing-c_1-req_1B-resp_1B_withMD-12       	   10000	    115799 ns/op
Unary-Tracing-c_1-req_1B-resp_1B_withBinMD-12    	   10000	    117684 ns/op
```

With the change:
```
Unary-Tracing-c_1-req_1B-resp_1B_noMD-12         	   10000	    103641 ns/op
Unary-Tracing-c_1-req_1B-resp_1B_withMD-12       	   10000	    105663 ns/op
Unary-Tracing-c_1-req_1B-resp_1B_withBinMD-12    	   10000	    109140 ns/op
```

For unary RPCs with metadata, latency is reduced about 10%.